% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_genoD.R
\docType{class}
\name{class_genoD-class}
\alias{class_genoD-class}
\alias{class_genoD}
\title{S4 class for marker genotype data with dominance}
\description{
S4 class for marker genotype data with dominance
}
\section{Slots}{

\describe{
\item{\code{ploidy}}{ploidy}

\item{\code{map}}{Marker map positions}

\item{\code{coeff}}{Coefficients of the additive marker effects (dim: indiv x marker)}

\item{\code{scale}}{Scaling factor between markers and indiv for additive effects}

\item{\code{G}}{Additive relationship matrix (from markers and potentially also pedigree)}

\item{\code{eigen.G}}{list of eigenvalues and eigenvectors for G}

\item{\code{coeff.D}}{coefficients of the dominance marker effects (dim: indiv x marker)}

\item{\code{scale.D}}{Scaling factor between markers and indiv for dominance effects}

\item{\code{D}}{Dominance relationship matrix}

\item{\code{eigen.D}}{list of eigenvalues and eigenvectors for D}

\item{\code{Fg}}{genomic inbreeding coefficient (based on dominance)}
}}

