% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_var.R
\docType{class}
\name{class_var-class}
\alias{class_var-class}
\alias{class_var}
\title{S4 class for variances}
\description{
S4 class for variances
}
\section{Slots}{

\describe{
\item{\code{add}}{additive}

\item{\code{g.resid}}{genetic residual}

\item{\code{resid}}{residual}

\item{\code{meanG}}{mean of diagonal of G}

\item{\code{meanOmega}}{mean of diagonal of Omega}

\item{\code{fixed.marker.cov}}{contribution of marker fixed effects to additive covariance between locations}
}}

