% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{corr}
\alias{corr}
\title{Trait correlations}
\usage{
corr(vars, traits = NULL, effect = NULL)
}
\arguments{
\item{vars}{object of \code{\link{class_var}} from \code{\link{Stage2}}}

\item{traits}{pair of traits}

\item{effect}{name of effect}
}
\value{
matrix
}
\description{
Trait correlations
}
\details{
Use either the argument \code{traits} or \code{effect}, not both. Using \code{traits} leads to a partitioning of the total correlation between those two traits, based on path analysis. Using \code{effect} displays the correlation between all traits for that effect. Use the \code{summary} command to see the names of the possible effects.
}
