% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_sum.R
\name{direct_sum}
\alias{direct_sum}
\title{Direct Sum for Symmetric Matrices}
\usage{
direct_sum(x)
}
\arguments{
\item{x}{list of matrices}
}
\value{
Sparse Matrix
}
\description{
Direct Sum for Symmetric Matrices
}
\details{
Computes the direct sum of the matrices in \code{x}
}
\keyword{internal}
