% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_mat.R
\name{G_mat}
\alias{G_mat}
\title{Additive genomic relationships}
\usage{
G_mat(geno, ploidy)
}
\arguments{
\item{geno}{Matrix of allele dosages (markers x indiv)}

\item{ploidy}{2 or 4}
}
\value{
List containing
\describe{
\item{coeff}{Coefficients of the marker effects (dim: indiv x marker)}
\item{scale}{Scaling factor between markers and indiv}
\item{mat}{G matrix}
}
}
\description{
Coefficients and relationship matrix for additive effects with bi-allelic markers
}
\details{
Additive effects are based on the traditional orthogonal decomposition of genetic variance in panmictic populations (Fisher 1918; Kempthorne 1957; Endelman et al. 2018). The G matrix is computed from the coefficients and scaling factor according to G = tcrossprod(coeff/scale). Missing genotype data is replaced with the population mean.
}
\references{
Fisher (1918) Trans. Roy. Soc. Edin. 52:399-433.

Kempthorne (1957) An Introduction to Genetic Statistics.

Endelman et al. (2018) Genetics 209:77-87.
}
