% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_threshold.R
\name{gwas_threshold}
\alias{gwas_threshold}
\title{Compute GWAS discovery threshold}
\usage{
gwas_threshold(geno, n.core = 1, alpha = 0.05)
}
\arguments{
\item{geno}{object of \code{\link{class_geno}}}

\item{n.core}{number of cores to use}

\item{alpha}{genome-wide significance level}
}
\value{
-log10(p) threshold
}
\description{
Compute GWAS discovery threshold
}
\details{
Uses a Bonferroni-type correction based on an effective number of markers that accounts for LD  (Moskvina and Schmidt, 2008).
}
\references{
Moskvina V, Schmidt KM (2008) On multiple-testing correction in genome-wide association studies. Genetic Epidemiology 32:567-573. doi:10.1002/gepi.20331
}
