% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbreeding.R
\name{inbreeding}
\alias{inbreeding}
\title{Genomic inbreeding coefficient}
\usage{
inbreeding(geno)
}
\arguments{
\item{geno}{object of \code{\link{class_geno}}}
}
\value{
data frame with F[G] and (when dominance is present) F[D]
}
\description{
Genomic inbreeding coefficient
}
\details{
Under the additive model, the inbreeding coefficient comes from the diagonal elements of the G matrix according to F = (G-1)/(ploidy-1). For dominance, the inbreeding coefficient is the scaled row-sum of the dominance coefficient matrix.
}
