% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Create Manhattan plot}
\usage{
manhattan_plot(data, chrom = NULL, thresh = NULL, rotate.label = FALSE)
}
\arguments{
\item{data}{data frame with columns for marker, chrom, position, and score}

\item{chrom}{optional, to plot only one chromosome}

\item{thresh}{optional, to include horizontal line at discovery threshold}

\item{rotate.label}{TRUE/FALSE whether to rotate x-axis labels to be perpendicular}
}
\value{
ggplot2 object
}
\description{
Create Manhattan plot
}
\details{
Assumes position in bp
}
