% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MME.R
\docType{class}
\name{MME-class}
\alias{MME-class}
\alias{MME}
\title{S4 class for solving the mixed model equations}
\description{
S4 class for solving the mixed model equations
}
\section{Slots}{

\describe{
\item{\code{data}}{data frame with id, env, blue, trait (optional)}

\item{\code{kernels}}{list of variance-covariance matrices for the genetic effects}

\item{\code{Rmat}}{residual variance-covariance matrix}
}}

