% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.geno.R
\name{predict}
\alias{predict}
\title{Predict individual values from marker effects}
\arguments{
\item{object}{object of \code{\link{class_geno}}}

\item{marker.effects}{data frame with columns "marker" and "effect"}
}
\value{
data frame with columns "id" and "value"
}
\description{
Predict individual values from marker effects
}
\details{
Use the \code{\link{blup}} function with what="AM" or "DM" to generate the data frame for \code{marker.effects}.
}
