% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_MME.R
\name{predict_MME}
\alias{predict_MME}
\title{Compute BLUPs by solving the Mixed Model Equations}
\usage{
predict_MME(data, weights = NULL, mask = NULL)
}
\arguments{
\item{data}{variable of class \code{\link{MME}}}

\item{weights}{named vector of weights for the genetic effects in BLUP. Default is 1 for all effects.}

\item{mask}{(optional) data frame with column "id" and optional columns "env", "trait"}
}
\value{
For single trait analysis, function returns a data frame with columns: id,blup,r2. For multi-trait analysis, a list is returned containing
\describe{
\item{blup}{data frame of blups}
\item{r2}{data frame of reliabilities}
}
}
\description{
Compute BLUPs by solving the Mixed Model Equations
}
\details{
Use the function \code{\link{Stage2}} to create the object of class \code{\link{MME}}. BLUPs are computed at the average value of the fixed effects. If \code{weights} is used, the names must exactly match the names of the kernels in \code{data}. Using the argument \code{mask}, the phenotypes for a subset of the population can be masked, to enable cross-validation.
}
