% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes.R
\name{print.Stage1}
\alias{print.Stage1}
\title{Print an object of class \code{Stage1}}
\usage{
\method{print}{Stage1}(x, ...)
}
\arguments{
\item{x}{An object fitted with the function \code{Stage1()}.}

\item{...}{Options used by the tibble package to format the output.}
}
\value{
an object inheriting from class \code{Stage1}.
}
\description{
Prints information about \code{Stage1} function.
}
\examples{
\dontrun{
model <- Stage1(
  filename = "pheno.csv",
  traits = "total.yield",
  solver = "spats", 
  spline = c("row", "range")
)
print(model)
}
}
