% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.geno.R
\name{quantile}
\alias{quantile}
\title{G matrix quantile}
\arguments{
\item{x}{object of \code{\link{class_geno}}}

\item{prob}{probability}
}
\value{
data frame with the quantile of the G matrix coefficients for each id
}
\description{
G matrix quantile
}
\details{
Unlike the S3 method, \code{prob} must have length = 1
}
