% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geno.R
\name{read_geno}
\alias{read_geno}
\title{Read marker genotype data}
\usage{
read_geno(
  filename,
  ploidy,
  map,
  min.minor.allele = 5,
  w = 1e-05,
  ped = NULL,
  dominance = FALSE,
  pop.file = NULL
)
}
\arguments{
\item{filename}{Name of CSV file with marker allele dosage}

\item{ploidy}{2,4,6,etc. (even numbers)}

\item{map}{TRUE/FALSE}

\item{min.minor.allele}{threshold for marker filtering (see Details)}

\item{w}{blending parameter (see Details)}

\item{ped}{optional, pedigree data frame with 3 or 4 columns (see Details)}

\item{dominance}{TRUE/FALSE whether to include dominance covariance (see Details)}

\item{pop.file}{CSV file defining populations}
}
\value{
Variable of class \code{\link{class_geno}}.
}
\description{
Read marker genotype data
}
\details{
When \code{map=TRUE}, first three columns of the file are marker, chrom, position. When \code{map=FALSE}, the first column is marker. Subsequent columns contain the allele dosage for individuals/clones, coded 0,1,2,...ploidy (fractional values are allowed). The input file for diploids can also be coded using {-1,0,1} (fractional values allowed). Additive coefficients are computed by subtracting the population mean from each marker, and the additive (genomic) relationship matrix is computed as G = tcrossprod(coeff)/scale. The scale parameter ensures the mean of the diagonal elements of G equals 1 under panmictic equilibrium. Missing genotype data is replaced with the population mean. 

Arguments \code{ped} and \code{w} can be used to blend the genomic G and pedigree A relationship matrices: Gw = (1-w)G + wA. The first three columns of \code{ped} are id, parent1, parent2. Missing parents must be coded NA. If a vector of w values is provided, the function returns a list of \code{\link{class_geno}} objects.

To include ungenotyped individuals in an H matrix, a fourth column is required in the pedigree data frame, which is an indicator variable (0/1) for which individuals to include. At present, this option cannot be combined with dominance. 
 
If the A matrix is not used, then G is blended with the identity matrix (times the mean diagonal of G) to improve numerical conditioning for matrix inversion. The default for w is 1e-5, which is somewhat arbitrary and based on tests with the vignette dataset. The D matrix is also blended with the identity matrix using 1e-5 for numerical conditioning.

When \code{dominance=FALSE}, non-additive effects are captured using a residual genetic effect, with zero covariance. If \code{dominance=TRUE}, a (digenic) dominance covariance matrix is used instead. 

The argument \code{min.minor.allele} specifies the minimum number of individuals that must contain the minor allele. Markers that do not meet this threshold are discarded.

Optional argument \code{pop.file} gives the name of a CSV file with two columns: id,pop. If the populations have different ploidy, this is indicated using a named vector for \code{ploidy}.
}
