% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_spatialtrend.R
\name{remove_spatialtrend}
\alias{remove_spatialtrend}
\title{Remove spatial trend}
\usage{
remove_spatialtrend(filename, traits, spline, effects = NULL)
}
\arguments{
\item{filename}{Name of CSV input file}

\item{traits}{trait names}

\item{spline}{vector of variable names for 2D spline with SpATS}

\item{effects}{data frame specifying other effects in the model (see Details)}
}
\value{
Data frame with adjusted phenotypes
}
\description{
Removes spatial trend to prepare for multi-trait Stage 1
}
\details{
SpATS used to remove 2D spatial trend for each field experiment, labeled with column 'expt' in the input file. Genotype labels are in column 'id' and modeled as i.i.d random effect. Argument \code{effects} is used to specify other i.i.d. effects besides genotype and has three columns: name, fixed, factor. The "name" column is a string that must match a column in the input file. The fixed column is a logical variable to indicate whether the effect is fixed (TRUE) or random (FALSE). The factor column is a logical variable to indicate whether the effect is a factor (TRUE) or numeric (FALSE). 

Argument \code{traits} is a character vector of trait names. Single-trait analyses are performed for each trait, and the results are combined in the output.
}
