% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes.R
\name{show,class_geno-method}
\alias{show,class_geno-method}
\title{Show an object of class \code{class_geno}}
\usage{
\S4method{show}{class_geno}(object)
}
\arguments{
\item{object}{An object of class \code{class_geno}.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Show an object of class \code{class_geno}
}
