% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes.R
\name{show,class_genoD-method}
\alias{show,class_genoD-method}
\title{Show an object of class \code{class_genoD}}
\usage{
\S4method{show}{class_genoD}(object)
}
\arguments{
\item{object}{An object of class \code{class_genoD}.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Show an object of class \code{class_genoD}
}
