% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage1.R
\name{Stage1}
\alias{Stage1}
\title{Stage 1 analysis of multi-environment trials}
\usage{
Stage1(
  filename,
  traits,
  effects = NULL,
  solver = "asreml",
  spline = NULL,
  silent = TRUE,
  workspace = c("500mb", "500mb")
)
}
\arguments{
\item{filename}{Name of CSV file}

\item{traits}{trait names (see Details)}

\item{effects}{data frame specifying other effects in the model (see Details)}

\item{solver}{one of the following: "asreml","spats"}

\item{spline}{vector of variable names for 2D spline with SpATS}

\item{silent}{TRUE/FALSE, whether to suppress REML output}

\item{workspace}{memory limits for ASRreml-R}
}
\value{
List containing
\describe{
\item{blue}{data frame of BLUEs for all environments}
\item{vcov}{list of variance-covariance matrices for the BLUEs, one per env}
\item{H2}{broad-sense H2 on a plot basis}
\item{resid}{list containing diagnostic plots and data frame of residuals}
\item{aic}{AIC for the BLUE model (only with asreml)}
}
}
\description{
Computes genotype BLUEs within each environment
}
\details{
The input file must have one column labeled "id" for the individuals and one labeled "env" for the environments. The data for each environment are analyzed independently with a linear mixed model. Although not used in Stage1, to include a genotype x location effect in \code{\link{Stage2}}, a column labeled "loc" should be present in the input file. 

Argument \code{effects} is used to specify other i.i.d. effects besides genotype and has three columns: name, fixed, factor. The "name" column is a string that must match a column in the input file. The fixed column is a logical variable to indicate whether the effect is fixed (TRUE) or random (FALSE). The factor column is a logical variable to indicate whether the effect is a factor (TRUE) or numeric (FALSE). 

Argument \code{solver} specifies which software to use for REML. Current options are "asreml" and "spats". For "spats", the argument \code{spline} must be a vector of length two, with the names of the x and y variables (respectively) for the 2D spline.

The heritability and residuals in the output are based on a random effects model for id.

Missing response values are omitted for single-trait analysis but retained for multi-trait analysis (unless both traits are missing), to allow for prediction in Stage 2. 

Argument \code{workspace} is a vector of length two containing the workspace and pworkspace limits for ASReml-R, with default values of 500mb. If you get an error about insufficient memory, try increasing the appropriate value (workspace for variance estimation and pworkspace for BLUE computation).

For multiple traits, only "asreml" is supported, and only the BLUE model is run, so the returned object does not contain H2.
}
