% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage2_prep.R
\name{Stage2_prep}
\alias{Stage2_prep}
\title{Prepare data for single-trait, Stage 2 analysis of multi-environment trials}
\usage{
Stage2_prep(blue.vcov, trait, exclude.id = character(0))
}
\arguments{
\item{blue.vcov}{list containing the blue.vcov matrix from Stage 1}

\item{trait}{trait name}

\item{exclude.id}{vector of individuals to exclude}
}
\value{
a list containing
\describe{
\item{blue}{data frame of BLUEs}
\item{Omega}{variance-covariance matrix of BLUEs}
}
}
\description{
Prepare data for single-trait, Stage 2 analysis of multi-environment trials
}
\details{
Designed to prepare data for \code{\link{Stage2}} based on output from \code{\link{Stage1}}. The argument \code{blue.vcov} is a list containing a matrix for each environment; the matrix contains BLUEs in column 1 and their variance-covariance in the remaining columns. The rownames are the id and env concatenated with ":".
}
