% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.geno.R
\name{summary.geno}
\alias{summary.geno}
\title{Summarize information about genomic relationships}
\arguments{
\item{object}{object of \code{\link{class_geno}}}
}
\value{
data frame
}
\description{
Summarize information about genomic relationships
}
\details{
First column contains diagonal elements of the G matrix. For objects with dominance, second column contains the genomic inbreeding coefficients (scaled row-sum of the D matrix).
}
