% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.var.R
\name{summary}
\alias{summary}
\title{Summarize variances and correlations}
\arguments{
\item{object}{object of \code{\link{class_var}}}
}
\value{
A matrix with the proportion of variance and covariance (for multi-trait). For multi-location or multi-trait analysis, a correlation matrix is also returned. If a G matrix was included, the additive correlation is above the diagonal, and genotypic correlation below.
}
\description{
Summarize variances and correlations
}
\details{
When reporting the partitioning of variance, the variance component for the additive effect is multiplied by the mean diagonal of the G matrix.
}
