% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniplot.R
\name{uniplot}
\alias{uniplot}
\title{Uniplot for multi-location models}
\usage{
uniplot(loadings, nudge = 0.1)
}
\arguments{
\item{loadings}{scaled factor loadings, from \code{\link{Stage2}}.}

\item{nudge}{distance to nudge labels}
}
\value{
ggplot2 object
}
\description{
Displays scaled loadings of the FA2 model
}
\details{
The squared radius for each point is the proportion of genetic variance explained by the latent factors. For points on the unit circle, the cosine of the subtended angle equals the correlation.
}
