% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheat-data.R
\docType{data}
\name{wheat.data}
\alias{wheat.data}
\title{Genomic prediction from secondary traits in wheat}
\format{
wheat.geno is object of \code{\link{class_geno}}. wheat.blues and wheat.vcov are output from \code{\link{Stage1}.}
}
\usage{
data(wheat)
}
\description{
Canopy temperature (CT) measurements collected during grain fill are used for genomic prediction of grain yield (GY) in wheat. Data come from the drought and extreme drought environments of Rutkoski et al. (2016).
}
\references{
Rutkoski et al. (2016) G3 (Bethesa) 6:2799–2808. https://doi.org/10.1534/g3.116.032888
}
\keyword{datasets}
