% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blup.R
\name{blup}
\alias{blup}
\title{BLUP}
\usage{
blup(data, geno = NULL, what, index.weights = NULL, gwas.ncore = 0L)
}
\arguments{
\item{data}{object of \code{\link{class_prep}} from \code{\link{blup_prep}}}

\item{geno}{object of \code{\link{class_geno}} from \code{\link{read_geno}}}

\item{what}{"id" or "marker"}

\item{index.weights}{named vector of index weights for the locations or traits}

\item{gwas.ncore}{Integer indicating number of cores to use for GWAS (default is 0 for no GWAS). Requires \code{what="markers"}.}
}
\value{
Data frames of BLUPs
}
\description{
BLUP
}
\details{
Argument \code{what="id"} leads to prediction of breeding values (BV) and genotypic values (GV), including the average fixed effect of the environments and any fixed effect markers.  For \code{what="marker"}, environment fixed effects are not included in the BLUP. Argument \code{index.weights} is a named vector (matching the names of the locations or traits), and the values are interpreted for standardized traits. An overall scaling factor is also applied so that the sum of the squared index coefficients equals 1.
}
