% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blup.R
\name{blup}
\alias{blup}
\title{BLUP}
\usage{
blup(data, geno = NULL, what, index.coeff = NULL, gwas.ncore = 0)
}
\arguments{
\item{data}{object of \code{\link{class_prep}} from \code{\link{blup_prep}}}

\item{geno}{object of \code{\link{class_geno}} from \code{\link{read_geno}}}

\item{what}{"id" or "marker"}

\item{index.coeff}{index coefficients for the locations or traits}

\item{gwas.ncore}{Number of cores to use for GWAS (default is 0 = no GWAS) if \code{what="markers"}}
}
\value{
Data frames of BLUPs
}
\description{
BLUP
}
\details{
Argument \code{index.coeff} should be a named vector, matching the names of the locations or traits. Index coefficients are normalized to have unit sum.
}
