% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blup_prep.R
\name{blup_prep}
\alias{blup_prep}
\title{Prepare data for BLUP}
\usage{
blup_prep(data, vcov = NULL, geno = NULL, vars, mask = NULL, method = NULL)
}
\arguments{
\item{data}{data frame of BLUEs from Stage 1}

\item{vcov}{list of variance-covariance matrices for the BLUEs}

\item{geno}{object of \code{\link{class_geno}} from \code{\link{read_geno}}}

\item{vars}{object of \code{\link{class_var}} from \code{\link{Stage2}}}

\item{mask}{(optional) data frame with possible columns "id","env","loc","trait"}

\item{method}{(optional) "MME", "Vinv", NULL (default). see Details}
}
\value{
Object of \code{\link{class_prep}}
}
\description{
Prepare data for BLUP
}
\details{
The \code{method} argument can be used to control how the linear system is solved. "MME" leads to inversion of the MME coefficient matrix, while "Vinv" leads to inversion of the overall var-cov matrix for the response vector. If NULL, the software uses whichever method involves inverting the smaller matrix. If the number of random effects (m) is less than the number of BLUEs (n), "MME" is used.

For the multi-location model, if all of the environments for a location are masked, the average of the other locations is used when computing average fixed effects.
}
