% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blup_prep.R
\name{blup_prep}
\alias{blup_prep}
\title{Prepare data for BLUP}
\usage{
blup_prep(data, vcov = NULL, geno = NULL, vars, mask = NULL)
}
\arguments{
\item{data}{data frame of BLUEs from Stage 1}

\item{vcov}{list of variance-covariance matrices for the BLUEs}

\item{geno}{object of \code{\link{class_geno}} from \code{\link{read_geno}}}

\item{vars}{object of \code{\link{class_var}} from \code{\link{Stage2}}}

\item{mask}{(optional) data frame with possible columns "id","env","trait"}
}
\value{
Object of \code{\link{class_prep}}
}
\description{
Prepare data for BLUP
}
