% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_geno.R
\docType{class}
\name{class_geno-class}
\alias{class_geno-class}
\alias{class_geno}
\title{S4 class for marker genotype data}
\description{
S4 class for marker genotype data
}
\section{Slots}{

\describe{
\item{\code{coeff}}{Coefficients of the marker effects (dim: indiv x marker)}

\item{\code{scale}}{Scaling factor between markers and indiv}

\item{\code{G}}{Additive (genomic) relationship matrix}

\item{\code{Ginv}}{Inverse G matrix}

\item{\code{map}}{Marker map positions}
}}

