% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_prep.R
\docType{class}
\name{class_prep-class}
\alias{class_prep-class}
\alias{class_prep}
\title{S4 class to prepare for blup}
\description{
S4 class to prepare for blup
}
\section{Slots}{

\describe{
\item{\code{y}}{Stage 1 BLUEs}

\item{\code{Z}}{Incidence matrix for random effects}

\item{\code{id}}{genotype identifiers}

\item{\code{var.u}}{variance of random effects}

\item{\code{Vinv}}{inverted covariance matrix of the Stage 1 BLUEs}

\item{\code{Pmat}}{P matrix from Searle}

\item{\code{fixed}}{fixed effect estimates}

\item{\code{random}}{random effect estimates}

\item{\code{add}}{matrix of additive variances from \code{\link{class_var}}}

\item{\code{loc.env}}{data frame with loc, env}

\item{\code{index.scale}}{sqrt of genetic variances for the locations/traits}

\item{\code{fixed.marker}}{names of fixed effect markers}
}}

