% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_prep.R
\docType{class}
\name{class_prep-class}
\alias{class_prep-class}
\alias{class_prep}
\title{S4 class to prepare for blup}
\description{
S4 class to prepare for blup
}
\section{Slots}{

\describe{
\item{\code{id}}{genotype identifiers}

\item{\code{ploidy}}{ploidy}

\item{\code{var.u}}{variance of random effects}

\item{\code{var.uhat}}{variance of BLUPs}

\item{\code{avg.env}}{average fixed effect of the environments}

\item{\code{heterosis}}{regression coefficients for inbreeding}

\item{\code{fixed.marker}}{fixed marker effects}

\item{\code{B}}{var-cov matrix for fixed effects}

\item{\code{random}}{random effect estimates}

\item{\code{add}}{matrix of additive variances from \code{\link{class_var}}}

\item{\code{g.iid}}{matrix of g.iid variances from \code{\link{class_var}}}

\item{\code{dom}}{matrix of dominance variances from \code{\link{class_var}}}
}}

