% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_var.R
\docType{class}
\name{class_var-class}
\alias{class_var-class}
\alias{class_var}
\title{S4 class for variances}
\description{
S4 class for variances
}
\section{Slots}{

\describe{
\item{\code{add}}{additive}

\item{\code{g.iid}}{iid genotype effect}

\item{\code{dom}}{dominance}

\item{\code{resid}}{residual}

\item{\code{diagG}}{average diagonal element of the G matrix}

\item{\code{diagD}}{average diagonal element of the D matrix}

\item{\code{vars}}{variances for reporting}

\item{\code{B}}{var-cov matrix of fixed effects for gain}

\item{\code{fix.eff.marker}}{names of fixed effect markers}
}}

