% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_var.R
\docType{class}
\name{class_var-class}
\alias{class_var-class}
\alias{class_var}
\title{S4 class for variances}
\description{
S4 class for variances
}
\section{Slots}{

\describe{
\item{\code{geno1}}{first genetic effect}

\item{\code{geno2}}{second genetic effect}

\item{\code{model}}{0=no markers, 1=add, 2=add+g.resid, 3=add+dom}

\item{\code{resid}}{residual}

\item{\code{vars}}{variances for reporting}

\item{\code{B}}{var-cov matrix of fixed effects for gain}

\item{\code{fix.eff.marker}}{names of fixed effect markers}
}}

