% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominance.R
\name{dominance}
\alias{dominance}
\title{Report dominance parameters}
\usage{
dominance(params, digits = 2)
}
\arguments{
\item{params}{list returned by \code{\link{Stage2}}}

\item{digits}{number of digits for rounding}
}
\value{
data frame with estimates and SE
}
\description{
Report dominance parameters
}
\details{
The dominance variance (Vd) and baseline heterosis (b) are quantified relative to additive variance (Va) and std. dev. (SDa), respectively. The estimate and SE of the ratios are calculated based on the delta method (Rice 2006, p. 162-166). Only implemented for single trait/location models.
}
\references{
Rice JA (2007) Mathematical Statistics and Data Analysis, 3rd ed. Duxbury, Pacific Grove.
}
