% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gain.R
\name{gain}
\alias{gain}
\title{Genetic gain}
\usage{
gain(
  input,
  merit = NULL,
  desired = NULL,
  restricted = NULL,
  traits = NULL,
  gamma = NULL,
  scale.coeff = NULL,
  solver = "ECOS",
  ...
)
}
\arguments{
\item{input}{variable of \code{\link{class_prep}} or list of matrices P,T,H (see Details)}

\item{merit}{named vector of merit coefficients, in genetic standard deviation units}

\item{desired}{named vector of desired gains, in genetic standard deviation units}

\item{restricted}{data frame of restricted traits, see Details}

\item{traits}{optional vector with exactly 2 trait names, to plot elliptical response}

\item{gamma}{contribution of non-additive values for genetic merit}

\item{scale.coeff}{should index coeff be scaled by genetic sd (see Details)}

\item{solver}{name of convex solver (default is "ECOS")}
}
\value{
List containing
\describe{
\item{matrices}{for input \code{\link{class_prep}}, a list with P,T,H}
\item{plot}{ellipse plot}
\item{table}{data frame with the response and index coefficients for all traits}
}
}
\description{
Genetic gain calculations
}
\details{
The standardized genetic response vector, \eqn{x=H^{-1}g}{x=H^-1 g}, is constrained to an ellipsoid with equation \eqn{x'Qx = i^2}{x'Qx=i^2}, where i is selection intensity and H is a diagonal matrix of genetic standard deviations. The matrix of the quadratic form \eqn{Q=U'U}{Q=U'U}, where \eqn{U=RT^{-1}H}{U=R(T^-1)H}; R is the upper triangular cholesky factor of P = variance-covariance matrix of the selection values (e.g., BLUPs or phenotypes); and T = covariance between the genetic and selection values. For multi-trait BLUPs T=P, while under phenotypic selection T is the genetic variance-covariance matrix.

The first argument \code{input} can be a list of three matrices (P,T,H) or a variable of class \code{\link{class_prep}}, in which case the matrices are computed internally. 

Either \code{merit} or \code{desired} can be used, not both. The former specifies the relative contribution of each trait to genetic merit, while the latter specifies the relative desired gain in genetic standard deviation units. All traits must be specified. Optional argument \code{restricted} is a data frame with columns "trait" and "sign", where the options for sign are "=",">","<", representing equal to zero, non-negative, and non-positive. When \code{desired} is used, the \code{restricted} argument is ignored.

The argument \code{gamma} controls the definition of genetic merit. (See notation in the journal publication.) The default is NULL, which implies breeding values. For purely additive values, use gamma = 0. For total genotypic value, use gamma = 1.

In the figure, the red dot is the intersection of the merit or desired gain vector with the ellipsoid, which is also the response under the desired gains index. The optimal merit response is shown with a blue dot.

In the function output, response values are in units of \eqn{i \sigma_g}{i sigma_g}, so multiplying by i gives the response for different intensities.

This function was originally designed to complement \code{\link{blup}}, which is based on an index of the form \eqn{I=b'BLUP[g]}{I=b'BLUP[g]}. For merit function \eqn{M=a'H^{-1}g}{M=a'(H^-1)g}, the index coefficients under BLUP are \eqn{b=H^{-1}a}{b=(H^-1)a}. If \eqn{c}{c} is the vector of coefficients passed to \code{\link{blup}}, internally the function uses \eqn{b=H^{-1}c}{b=(H^-1)c}. In other words, the argument should be scaled like \eqn{a}{a}. To return index coefficients on this scale, use \code{scale.coeff}=TRUE. To return the unscaled coefficients, use \code{scale.coeff}=FALSE. The default value, NULL, will use TRUE when the input is of class \code{\link{class_prep}} and FALSE when the input is a list of three matrices.
}
