% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gain.R
\name{gain}
\alias{gain}
\title{Genetic gain}
\usage{
gain(input, traits = NULL, coeff = NULL, restricted = NULL, solver = "ECOS")
}
\arguments{
\item{input}{either object of \code{\link{class_prep}} or quad.mat returned by this function}

\item{traits}{optional, plots ellipse tradeoff}

\item{coeff}{optional, index coefficients expressed in genetic standard deviation units}

\item{restricted}{data frame of restricted traits see Details}

\item{solver}{name of convex solver (default is "ECOS")}
}
\value{
List containing
\describe{
\item{quad.mat}{quadratic matrix for the ellipsoid}
\item{plot}{ellipse plot}
\item{table}{data frame with gain and coefficients for the traits}
}
}
\description{
Genetic gain for breeding values
}
\details{
Optional argument \code{restricted} is a data frame with columns "trait" and "sign", where the options for sign are "=",">","<", representing equal to zero, non-negative, and non-positive.
}
