% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.geno.R
\name{predict}
\alias{predict}
\title{Predict additive (breeding) values from marker effects}
\arguments{
\item{object}{object of \code{\link{class_geno}}}

\item{marker.effects}{data frame with columns "marker","add.effect"}
}
\value{
data frame with columns "id", "BV"
}
\description{
Predict additive (breeding) values from marker effects
}
\details{
Use the \code{\link{blup}} function with \code{what="marker"} to generate the data frame for \code{marker.effects}.
}
