% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geno.R
\name{read_geno}
\alias{read_geno}
\title{Read marker genotype data}
\usage{
read_geno(filename, ploidy, map, eigen.tol = 1e-09)
}
\arguments{
\item{filename}{Name of CSV file}

\item{ploidy}{2,4,6,etc. (even numbers)}

\item{map}{TRUE/FALSE}

\item{eigen.tol}{1e-9}
}
\value{
Variable of class \code{\link{class_geno}}.
}
\description{
Read marker genotype data
}
\details{
When \code{map=TRUE}, first three columns of the file are marker, chrom, position. When \code{map=FALSE}, the first column is marker. Subsequent columns contain the allele dosage for individuals/clones, coded 0,1,2,...ploidy (fractional values are allowed). The input file for diploids can also be coded using {-1,0,1} (fractional values allowed). Additive coefficients are computed by subtracting the population mean from each marker, and the additive (genomic) relationship matrix is computed as G = tcrossprod(coeff)/scale. The scale parameter ensures the mean of the diagonal elements of G equals 1 under panmictic equilibrium. Missing genotype data is replaced with the population mean. For numerical conditioning, eigenvalues of G smaller than \code{eigen.tol} are replaced by \code{eigen.tol}.
}
