% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage1.R
\name{Stage1}
\alias{Stage1}
\title{Stage 1 analysis of multi-environment trials}
\usage{
Stage1(
  data,
  traits,
  effects = NULL,
  silent = TRUE,
  workspace = "500mb",
  pworkspace = "500mb"
)
}
\arguments{
\item{data}{input data frame (see Details)}

\item{traits}{trait names (see Details)}

\item{effects}{data frame specifying other effects in the model (see Details)}

\item{silent}{TRUE/FALSE, whether to suppress ASReml-R output}

\item{workspace}{memory limit for ASRreml-R variance estimation}

\item{pworkspace}{memory limit for ASRreml-R BLUE computation}
}
\value{
List containing
\describe{
\item{blue}{data frame of BLUEs for all environments}
\item{vcov}{list of variance-covariance matrices for the BLUEs, one per env}
\item{H2}{broad-sense H2 on a plot basis (only for single trait)}
\item{resid}{list containing boxplot, qqplot, and table of residuals (only for single trait)}
}
}
\description{
Computes genotype BLUEs within each environment using ASReml-R
}
\details{
The variable \code{data} must have one column labeled "id" for the individuals and one labeled "env" for the environments. The data for each environment are analyzed independently with a linear mixed model. Although not used in Stage1, to include a genotype x location effect in \code{\link{Stage2}}, a column labeled "loc" should be included in \code{data}. 

Including multiple traits in \code{trait} triggers a multivariate analysis, but for computational reasons, only 2 traits are analyzed at a time. With more than 2 traits, the software analyzes all pairs of traits. For single trait analysis, broad-sense H2 on a plot basis is computed from the variance components for each env, with genotype as a random effect. The residuals from this analysis are also returned as a table and plots.

Argument \code{effects} is used to specify other i.i.d. effects besides genotype and has three columns: name, fixed, factor. The "name" column is a string that must match a column in \code{data}. The fixed column is a logical variable to indicate whether the effect is fixed (TRUE) or random (FALSE). The factor column is a logical variable to indicate whether the effect is a factor (TRUE) or numeric (FALSE). 

Missing response values are omitted for single-trait analysis but retained for multi-trait analysis (unless both traits are missing), to allow for prediction in Stage 2. By default, the workspace and pworkspace limits for ASReml-R are set at 500mb. If you get an error about insufficient memory, try increasing the appropriate value (workspace for variance estimation and pworkspace for BLUE computation).
}
\references{
Damesa et al. 2017. Agronomy Journal 109: 845-857. doi:10.2134/agronj2016.07.0395
}
