% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage1.R
\name{Stage1}
\alias{Stage1}
\title{Stage 1 analysis of multi-environment trials}
\usage{
Stage1(
  filename,
  traits,
  effects = NULL,
  solver = "asreml",
  spline = NULL,
  silent = TRUE,
  workspace = c("500mb", "500mb"),
  max.iter = 30,
  env = NULL
)
}
\arguments{
\item{filename}{Name of CSV file}

\item{traits}{trait names (see Details)}

\item{effects}{data frame specifying other effects in the model (see Details)}

\item{solver}{one of the following: "asreml","spats"}

\item{spline}{vector of variable names for 2D spline with SpATS}

\item{silent}{TRUE/FALSE, whether to suppress REML output}

\item{workspace}{memory limits for ASRreml-R}

\item{max.iter}{maximum number of iterations for ASRreml-R}

\item{env}{character vector of environments to analyze (default is all)}
}
\value{
List containing
\describe{
\item{blues}{data frame of BLUEs}
\item{vcov}{list of variance-covariance matrices for the BLUEs, one per env}
\item{fit}{data frame with broad-sense H2 and/or AIC}
\item{resid}{For single trait, list of diagnostic plots and data frame of residuals. For multi-trait, list of resid var-cov matrices.}
}
}
\description{
Computes genotype BLUEs for each experiment
}
\details{
The input file must have one column labeled "id" for the individuals and one labeled "env" for the environments. The data for each environment are analyzed independently with a linear mixed model. Although not used in Stage1, to include a genotype x location effect in \code{\link{Stage2}}, a column labeled "loc" should be present in the input file. 

Argument \code{effects} is used to specify other i.i.d. effects besides genotype and has three columns: name, fixed, factor. The "name" column is a string that must match a column in the input file. The fixed column is a logical variable to indicate whether the effect is fixed (TRUE) or random (FALSE). The factor column is a logical variable to indicate whether the effect is a factor (TRUE) or numeric (FALSE). 

Argument \code{solver} specifies which software to use for REML. Current options are "asreml" and "spats". For "spats", the argument \code{spline} must be a vector of length two, with the names of the x and y variables (respectively) for the 2D spline.

The heritability and residuals in the output are based on a random effects model for id. As of v1.13, separate values for plot and entry-mean H2 in the broad-sense are calculated. With asreml, plot-H2 is from the variance components, while entry-H2 is the average reliability from \code{\link{blup}}. SpATS reports a "generalized" H2, which is equivalent to average reliability.

Missing response values are omitted for single-trait analysis but retained for multi-trait analysis (unless both traits are missing), to allow for prediction in Stage 2. 

Argument \code{workspace} is a vector of length two containing the workspace and pworkspace limits for ASReml-R, with default values of 500mb. If you get an error about insufficient memory, try increasing the appropriate value (workspace for variance estimation and pworkspace for BLUE computation).

For multiple traits, only "asreml" is supported, and only the BLUE model is run, so the returned object does not contain H2. 

If the input file has a column "expt", this allows for the use of separate spatial models for multiple experiments within an environment (only for single trait): each experiment is first analyzed separately, and then the BLUEs from all experiments per env are jointly analyzed to compute a single BLUE per env. The estimation errors from each experiment are propagated into the multi-expt model using ASReml-R. The situation is different with multi-trait analysis, as all experiments are analyzed jointly per env, with a fixed effect for expt but a common residual model. Any additional cofactors (e.g., block) that are nested within expt need to be explicitly nested! 

Argument \code{env} can be used to limit the analysis to only certain environments in the input file. This is useful when different models are needed in different env.
}
