% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage2.R
\name{Stage2}
\alias{Stage2}
\title{Stage 2 analysis of multi-environment trials}
\usage{
Stage2(
  data,
  vcov = NULL,
  geno = NULL,
  fix.eff.marker = NULL,
  silent = TRUE,
  workspace = "500mb",
  non.add = "g.resid",
  max.iter = 20,
  covariates = NULL
)
}
\arguments{
\item{data}{data frame of BLUEs from Stage 1 (see Details)}

\item{vcov}{named list of variance-covariance matrices for the BLUEs}

\item{geno}{output from \code{\link{read_geno}}}

\item{fix.eff.marker}{markers in \code{geno} to include as additive fixed effect covariates}

\item{silent}{TRUE/FALSE, whether to suppress ASReml-R output}

\item{workspace}{Memory limit for ASRreml-R variance estimation}

\item{non.add}{one of the following: "none","g.resid","dom"}

\item{max.iter}{maximum number of iterations for asreml}

\item{covariates}{names of other covariates in \code{data}}
}
\value{
List containing
\describe{
\item{aic}{AIC}
\item{vars}{variance components for \code{\link{blup_prep}}, as variable of class \code{\link{class_var}}}
\item{params}{Estimates and SE for fixed effects and variance components}
\item{random}{Random effect predictions}
\item{loadings}{scaled loadings for the FA2 multi-loc model}
}
}
\description{
Stage 2 analysis of multi-environment trials
}
\details{
Stage 2 of the two-stage approach described by Damesa et al. 2017, using ASReml-R for variance component estimation. The variable \code{data} has three mandatory column: id, env, BLUE. Optionally, \code{data} can have a column labeled "loc", which changes the main effect for genotype into a separable genotype-within-location effect, using a FA2 covariance model for the locations. Optionally, \code{data} can have a column labeled "trait", which uses an unstructured covariance model. The multi-location and multi-trait analyses cannot be combined. Missing data are allowed in the multi-trait but not the single-trait analysis. The argument \code{geno} is used to partition genetic values into additive and non-additive components. Any individuals in \code{data} that are not present in \code{geno} are discarded. 

The argument \code{vcov} is used to partition the macro- and micro-environmental variation, which are called GxE and residual in the output. \code{vcov} is a named list of variance-covariance matrices for the BLUEs within each environment, with id for rownames (single trait) or id:trait. The order in \code{vcov} and \code{data} should match. Both \code{data} and \code{vcov} can be created using the function \code{\link{Stage1}}. 

Because ASReml-R can only use relationship matrices defined in the global environment, this function creates and then removes global variables when either \code{vcov} or \code{geno} is used. By default, the workspace memory for ASReml-R is set at 500mb. If you get an error about insufficient memory, try increasing it. ASReml-R version 4.1.0.148 or later is required. 

The \code{covariates} option is only available for single trait/loc analysis.
}
\references{
Damesa et al. 2017. Agronomy Journal 109: 845-857. doi:10.2134/agronj2016.07.0395
}
