% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.geno.R
\name{summary.geno}
\alias{summary.geno}
\title{Summarize information about genomic relationships}
\arguments{
\item{object}{object of \code{\link{class_geno}}}
}
\value{
matrix
}
\description{
Summarize information about genomic relationships
}
\details{
First column contains diagonal elements of the G matrix. For objects with dominance, column 2 is diagonal elements of the D matrix, and column 3 is the genomic inbreeding coefficient (scaled row-sum of the D matrix).
}
