% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.var.R
\name{summary}
\alias{summary}
\title{Summarize variances and correlations}
\arguments{
\item{object}{object of \code{\link{class_var}}}
}
\value{
For univariate analysis, a matrix with the proportion of variance. For multi-location or multi-trait analysis, a list containing the proportion of variance and the correlation matrix.
}
\description{
Summarize variances and correlations
}
\details{
When reporting the partitioning of variance, the variance component for the additive effect is multiplied by the mean diagonal of the G matrix. When a G matrix has been included, the correlation matrix shown is for the additive value (plus any markers).
}
