% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.var.R
\name{summary}
\alias{summary}
\title{Summarize variances and correlations}
\arguments{
\item{object}{object of \code{\link{class_var}}}
}
\value{
For univariate analysis, a matrix of variances. For multi-location or multi-trait analysis, a list containing the variance matrix and the correlation matrix.
}
\description{
Summarize variances and correlations
}
\details{
When reporting the partitioning of variance, the variance component for the additive effect is multiplied by the mean diagonal of the G matrix.
}
