% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.var.R
\name{summary}
\alias{summary}
\title{Summarize variances and correlations}
\arguments{
\item{object}{object of \code{\link{class_var}}}
}
\value{
For univariate analysis, a matrix with the proportion of variance (R2). For multi-location or multi-trait analysis, a correlation matrix is also returned. If a G matrix was included in the analysis, the correlation is between additive values; otherwise, it is between genotypic values. If fixed effect markers are present, the additive correlation is above the diagonal, and below the diagonal is the proportion of additive covariance due to the fixed effect markers.
}
\description{
Summarize variances and correlations
}
\details{
When reporting the partitioning of variance, the variance component for the additive effect is multiplied by the mean diagonal of the G matrix.
}
